requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
tqdm>=4.65.0
lpips>=0.1.4
swanlab>=0.3.0
einops>=0.7.0
pyyaml>=6.0

# Optional: for Mamba support
# mamba-ssm>=1.0.0

# For distributed training
# ninja>=1.11.0
